requirements.txt
asttokens==2.2.1
audiosegment==0.23.0
backcall==0.2.0
cmake==3.26.3
decorator==5.1.1
dill==0.3.6
executing==1.2.0
filelock==3.11.0
future==0.18.3
ipython==8.12.0
jedi==0.18.2
Jinja2==3.1.2
lit==16.0.1
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mpmath==1.3.0
networkx==3.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
pydub==0.25.1
pygame==2.3.0
Pygments==2.15.0
PySide6==6.5.0
PySide6-Addons==6.5.0
PySide6-Essentials==6.5.0
python-vlc==3.0.18121
shiboken6==6.5.0
six==1.16.0
stack-data==0.6.2
sympy==1.11.1
torch==2.0.0
traitlets==5.9.0
triton==2.0.0
typing_extensions==4.5.0
wcwidth==0.2.6
webrtcvad==2.0.10
